# contracts/openapi.json
{
  "components": {
    "schemas": {
      "ApiError": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "error"
        ],
        "title": "ApiError",
        "type": "object"
      },
      "BriefingItemResponse": {
        "additionalProperties": false,
        "properties": {
          "estimated_seconds": {
            "title": "Estimated Seconds",
            "type": "integer"
          },
          "hotspot_id": {
            "format": "uuid",
            "title": "Hotspot Id",
            "type": "string"
          },
          "position": {
            "title": "Position",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "tts_text": {
            "title": "Tts Text",
            "type": "string"
          }
        },
        "required": [
          "position",
          "hotspot_id",
          "title",
          "tts_text",
          "estimated_seconds"
        ],
        "title": "BriefingItemResponse",
        "type": "object"
      },
      "BriefingResponse": {
        "additionalProperties": false,
        "properties": {
          "date": {
            "format": "date",
            "title": "Date",
            "type": "string"
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "generated_at": {
            "format": "date-time",
            "title": "Generated At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "intro_text": {
            "title": "Intro Text",
            "type": "string"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/BriefingItemResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "outro_text": {
            "title": "Outro Text",
            "type": "string"
          },
          "timezone": {
            "title": "Timezone",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "date",
          "timezone",
          "version",
          "generated_at",
          "expires_at",
          "intro_text",
          "items",
          "outro_text"
        ],
        "title": "BriefingResponse",
        "type": "object"
      },
...
      "DeepAnalysis": {
        "additionalProperties": false,
        "properties": {
          "background": {
            "title": "Background",
            "type": "string"
          },
          "exam_relevance": {
            "title": "Exam Relevance",
            "type": "string"
          },
          "possible_angles": {
            "items": {
              "type": "string"
            },
            "title": "Possible Angles",
            "type": "array"
          },
          "what_changed": {
            "title": "What Changed",
            "type": "string"
          },
          "why_it_matters": {
            "title": "Why It Matters",
            "type": "string"
          }
        },
        "required": [
          "background",
          "what_changed",
          "why_it_matters",
          "exam_relevance",
          "possible_angles"
        ],
        "title": "DeepAnalysis",
        "type": "object"
      },
      "ErrorDetail": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "request_id": {
            "format": "uuid",
            "title": "Request Id",
            "type": "string"
          },
          "retryable": {
            "title": "Retryable",
            "type": "boolean"
          }
        },
        "required": [
          "code",
          "message",
          "request_id",
          "retryable"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
...
      "HotspotDetail": {
        "additionalProperties": false,
        "properties": {
          "brief": {
            "title": "Brief",
            "type": "string"
          },
          "category": {
            "title": "Category",
            "type": "string"
          },
          "content_version": {
            "title": "Content Version",
            "type": "integer"
          },
          "deep_analysis": {
            "$ref": "#/components/schemas/DeepAnalysis"
          },
          "evidence_refs": {
            "items": {
              "type": "string"
            },
            "title": "Evidence Refs",
            "type": "array"
          },
          "exam_types": {
            "items": {
              "type": "string"
            },
            "title": "Exam Types",
            "type": "array"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "importance": {
            "enum": [
              "core",
              "other"
            ],
            "title": "Importance",
            "type": "string"
          },
          "insufficient_evidence": {
            "title": "Insufficient Evidence",
            "type": "boolean"
          },
          "key_points": {
            "items": {
              "type": "string"
            },
            "title": "Key Points",
            "type": "array"
          },
          "knowledge_refs": {
            "items": {
              "type": "string"
            },
            "title": "Knowledge Refs",
            "type": "array"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/SourceLink"
            },
            "title": "Sources",
            "type": "array"
          },
          "subcategories": {
            "items": {
              "type": "string"
            },
            "title": "Subcategories",
            "type": "array"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "category",
          "subcategories",
          "exam_types",
          "importance",
          "brief",
          "summary",
          "deep_analysis",
          "key_points",
          "knowledge_refs",
          "evidence_refs",
          "insufficient_evidence",
          "content_version",
          "sources"
        ],
        "title": "HotspotDetail",
        "type": "object"
      },
      "HotspotItem": {
        "additionalProperties": false,
        "properties": {
          "category": {
            "title": "Category",
            "type": "string"
          },
          "content_version": {
            "title": "Content Version",
            "type": "integer"
          },
          "exam_types": {
            "items": {
              "type": "string"
            },
            "title": "Exam Types",
            "type": "array"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "importance": {
            "enum": [
              "core",
              "other"
            ],
            "title": "Importance",
            "type": "string"
          },
          "published_at": {
            "format": "date-time",
            "title": "Published At",
            "type": "string"
          },
          "source_names": {
            "items": {
              "type": "string"
            },
            "title": "Source Names",
            "type": "array"
          },
          "subcategories": {
            "items": {
              "type": "string"
            },
            "title": "Subcategories",
            "type": "array"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "summary",
          "category",
          "subcategories",
          "exam_types",
          "importance",
          "published_at",
          "source_names",
          "content_version"
        ],
        "title": "HotspotItem",
        "type": "object"
      },
      "HotspotList": {
        "additionalProperties": false,
        "properties": {
          "date": {
            "format": "date",
            "title": "Date",
            "type": "string"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/HotspotItem"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "date",
          "items",
          "next_cursor"
        ],
        "title": "HotspotList",
        "type": "object"
      },
      "SourceLink": {
        "additionalProperties": false,
        "properties": {
          "name": {
            "title": "Name",
            "type": "string"
          },
          "published_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published At"
          },
          "url": {
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "name",
          "url",
          "published_at"
        ],
        "title": "SourceLink",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "公考晨报 API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/briefings/today": {
      "get": {
        "operationId": "todays_briefing_v1_briefings_today_get",
        "parameters": [
          {
            "in": "query",
            "name": "date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date"
            }
          },
          {
            "in": "query",
            "name": "timezone",
            "required": false,
            "schema": {
              "default": "Asia/Shanghai",
              "title": "Timezone",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "locale",
            "required": false,
            "schema": {
              "default": "zh-CN",
              "title": "Locale",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "if-none-match",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "If-None-Match"
            }
          },
          {
            "in": "header",
            "name": "x-client-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Client-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BriefingResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "summary": "Todays Briefing"
      }
    },
...
    "/v1/hotspots": {
      "get": {
        "operationId": "list_hotspots_v1_hotspots_get",
        "parameters": [
          {
            "in": "query",
            "name": "date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date"
            }
          },
          {
            "in": "query",
            "name": "timezone",
            "required": false,
            "schema": {
              "default": "Asia/Shanghai",
              "title": "Timezone",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "category",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 50,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "if-none-match",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "If-None-Match"
            }
          },
          {
            "in": "header",
            "name": "x-client-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Client-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HotspotList"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Unauthorized"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "summary": "List Hotspots"
      }
    },
    "/v1/hotspots/{hotspot_id}": {
      "get": {
        "operationId": "hotspot_detail_v1_hotspots__hotspot_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "hotspot_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Hotspot Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "if-none-match",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "If-None-Match"
            }
          },
          {
            "in": "header",
            "name": "x-client-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Client-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HotspotDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "summary": "Hotspot Detail"
      }
    }
  }
}